#27 refactor: Fix stock handling errors and ensure proper functionality
