--- src/main/java/com/infinityiterators/shipment/model/OutRecordDTO.java
@@ -1,23 +1,0 @@
-package com.infinityiterators.shipment.model;
-
-
-import lombok.*;
-
-import java.util.Date;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-public class OutRecordDTO {
-
-    private String ShipmentID;
-    private int OutAmount;
-    private Date ShipmentDate;
-    private String BookID;
-
-
-
-
-}
