--- src/main/java/com/infinityiterators/shipment/model/ShipmentDAO.java
@@ -1,11 +1,0 @@
-package com.infinityiterators.shipment.model;
-
-import com.infinityiterators.shipment.controller.PrintResult;
-import com.infinityiterators.shipment.service.ShipmentService;
-
-import java.util.List;
-import java.util.Scanner;
-
-public class ShipmentDAO {
-
-}
