--- src/main/java/com/infinityiterators/shipment/service/shipmentMapper.java
@@ -1,4 +1,0 @@
-package com.infinityiterators.shipment.service;
-
-public class shipmentMapper {
-}
