+++ src/main/resources/com/infinityiterators/bookwms/receipt/mapper/ReceiptMapper.xml
@@ -3,16 +3,16 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com/infinityiterators/receipt/mapper/ReceiptMapper.java">
-    <select id="selectAllStock" resultType="com/infinityiterators/receipt/model/dto/BookDTO.java">
-        SELECT BOOK_ID,
-        TITLE,
-        AUTHOR,
-        PUBLISHER,
+<mapper namespace="com.infinityiterators.bookwms.receipt.mapper.ReceiptMapper">
+    <select id="selectAllBook" resultType="com.infinityiterators.bookwms.receipt.model.dto.BookDTO">
+        SELECT BookID,
+        Title,
+        Author,
+        Publisher,
         ISBN
 <!--        NATIONCODE,-->
 <!--        GENRECODE-->
-        FROM tbl_book
+        FROM Book
         ORDER BY TITLE;
     </select>
 
@@ -30,9 +30,9 @@
         )
     </insert>
 
-    <update id="uadateBook" parameterType="com/infinityiterators/receipt/model/dto/StockDTO">
-        UPDATE STOCK
-        SET BOOKID = #{bookID}
-        , BOOKAMOUNT = #{inAmount}
+    <update id="updateBook" parameterType="com.infinityiterators.bookwms.receipt.model.dto.StockDTO">
+        UPDATE Stock
+        SET BookID = #{bookID}
+        , BookAmount = #{inAmount}
     </update>
 </mapper>
